requirements.txt
alembic==0.8.8
blinker==1.4
click==6.6
Flask==0.11.1
Flask-Admin==1.4.2
Flask-Mail==0.9.1
Flask-Markdown==0.3
Flask-Migrate==2.0.0
Flask-Script==2.0.5
Flask-SQLAlchemy==2.1
Flask-Uploads==0.2.1
Flask-WTF==0.13.1
gunicorn==19.6.0
itsdangerous==0.24
Jinja2==2.8
Mako==1.0.4
Markdown==2.6.7
MarkupSafe==0.23
mpmath==0.19
numpy==1.11.2
pandas==0.19.0
psycopg2==2.6.2
py-bcrypt==0.4
python-dateutil==2.5.3
python-editor==1.0.1
python-slugify==1.2.1
pytz==2016.7
scipy==0.18.1
six==1.10.0
SQLAlchemy==1.1.2
sympy==1.0
Unidecode==0.4.19
Werkzeug==0.11.11
WTForms==2.1
